+++ test/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_11_03_083443) do
+ActiveRecord::Schema[7.0].define(version: 2022_11_14_091221) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
   enable_extension "unaccent"
@@ -105,7 +105,6 @@ ActiveRecord::Schema[7.0].define(version: 2022_11_03_083443) do
     t.datetime "updated_at", precision: nil, null: false
     t.string "first_name"
     t.string "last_name"
-    t.string "role"
     t.boolean "is_active", default: true
     t.string "invitation_token"
     t.datetime "invitation_created_at", precision: nil
@@ -118,6 +117,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_11_03_083443) do
     t.string "crossdomain_devise_token"
     t.datetime "crossdomain_devise_set_at"
     t.string "sign_out_salt_part"
+    t.jsonb "roles", default: []
     t.index ["crossdomain_devise_token"], name: "index_folio_accounts_on_crossdomain_devise_token"
     t.index ["email"], name: "index_folio_accounts_on_email", unique: true
     t.index ["invitation_token"], name: "index_folio_accounts_on_invitation_token", unique: true